# api.yaml
swagger: "2.0.0"
info:
  title: ShopOnAir
  description: Api documentation for the shoponair website
  version: 1.0.0
  contact:
    email: choudharysidharth082000@gmail.com
  servers:
    - url: "http://localhost:3000"
      description: "Local Server Running"
    - url: "https://shoponairbackend.herokuapp.com"
      description: "Main Server"
schemes: [http, https]
paths:
...
  /v1/admin/blockUser/{userID}:
    put: 
      tags:
        - name: Admin
      name: Block/Unblock User
      description: Block/Unblock the user 
      consumes:
        multipart/formdata
      produces:
        application/json
      parameters:
        - name: userID
          in: path
          type: string
          required: true
        - name: status
          in: formData
          type: boolean
          required: true
      responses:
        200:
          description: OK
...
  /v1/website/getWebsites/{offset}/{limit}:
    get:
      tags:
        - name: Website
      name: Getting all the websites
      description: Getting all the websites in the main admin page
      parameters:
        - name: offset
          in: path
          type: string
          required: true
        - name: limit
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
...
  /v1/website/getWebsiteByUser/{userID}/{offset}/{limit}:
    get:
      tags:
        - name: Website
      name: Getting all the websites of particular user
      description: Getting all the websites in the main admin page
      parameters:
        - name: userID
          in: path
          type: string
          required: true
        - name: offset
          in: path
          type: string
          required: true
        - name: limit
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
...
  /v1/admin/deleteUser/{userID}:
    delete:
      tags:
        - name: Admin
      name: Delete User
      description: Deleting the user
      parameters:
        - name: userID
          in: path 
          required: true
          type: string
      responses:
        200:
          description: OK
  /v1/website/deleteWebsite/{websiteID}:
    delete:
        tags:
          - name: Website
        name: Deleting Website
        description: Deleting the website for the database
        parameters:
          - name: websiteID
            in: path
            type: string
            required: true
        responses:
          200:
            description: OK
  /v1/website/getWebsiteByID/{websiteID}:
    get:
      tags:
        - name: Website
      name: Get Website By Website Website ID
      description: Getting the website
      parameters:
        - name: websiteID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
...
  /v1/website/addWebsite/{userID}:
    post:
      tags:
        - name: Website
      name: Creating Website
      description: This api will create the website for the User
      consumes:
        multipart/formdata
      produces:
        application/json
      parameters:
        - name: userID
          in: path
          type: string
          required: true
        - name: siteName
          in: formData
          type: string
          required: true
        - name: siteURL
          in: formData
          type: string
          required: true
        - name: consumerKey
          in: formData
          type: string
          required: true
        - name: consumerSecret
          in: formData
          type: string
          required: true
        - name: discription
          in: formData
          type: string
          required: true
      responses:
        200:
          description: OK
  /v1/website/editWebsite/{userID}/{websiteID}:
    put:
      tags:
        - name: Website
      name: Edting Website
      description: This api will edit the website for the User
      consumes:
        multipart/formdata
      produces:
        application/json
      parameters:
        - name: userID
          in: path
          type: string
          required: true
        - name: websiteID
          in: path
          type: string
          required: true
        - name: siteName
          in: formData
          type: string
          required: true
        - name: siteURL
          in: formData
          type: string
          required: true
        - name: consumerKey
          in: formData
          type: string
          required: true
        - name: consumerSecret
          in: formData
          type: string
          required: true
        - name: discription
          in: formData
          type: string
          required: true
      responses:
        200:
          description: OK
...